scripts/requirements.txt
jobspy>=1.1.0
supabase>=2.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
